pyproject.toml
[project]
name = "langgraph-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.1.0",
    "fpdf>=1.7.2",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.3.30",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langgraph-cli[inmem]>=0.4.3",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "plotly>=5.24.0",
    "pandas>=2.2.0",
]
